move migration related dbaccessors to migration util file
