+++ corehq/motech/repeaters/dbaccessors.py
@@ -227,34 +227,6 @@ def get_sql_repeat_records_by_payload_id(domain, payload_id):
             .all())
 
 
-def get_repeaters_by_domain(domain):
-    from .models import Repeater
-
-    results = Repeater.get_db().view('repeaters/repeaters',
-        startkey=[domain],
-        endkey=[domain, {}],
-        include_docs=True,
-        reduce=False,
-    ).all()
-
-    return [Repeater.wrap(result['doc']) for result in results
-            if Repeater.get_class_from_doc_type(result['doc']['doc_type'])
-            ]
-
-
-def _get_repeater_ids_by_domain(domain):
-    from .models import Repeater
-
-    results = Repeater.get_db().view('repeaters/repeaters',
-        startkey=[domain],
-        endkey=[domain, {}],
-        include_docs=False,
-        reduce=False,
-    ).all()
-
-    return [result['id'] for result in results]
-
-
 def iterate_repeat_records_for_ids(doc_ids):
     from .models import RepeatRecord
     return (RepeatRecord.wrap(doc) for doc in iter_docs(RepeatRecord.get_db(), doc_ids))
@@ -308,30 +280,3 @@ def delete_all_repeaters():
     from .models import Repeater
     for repeater in Repeater.get_db().view('repeaters/repeaters', reduce=False, include_docs=True).all():
         Repeater.wrap(repeater['doc']).delete()
-
-
-def get_all_repeater_docs():
-    from .models import Repeater
-    results = Repeater.get_db().view('repeaters/repeaters', reduce=False, include_docs=True).all()
-    return [
-        repeater['doc'] for repeater in results
-        if Repeater.get_class_from_doc_type(repeater['doc']['doc_type'])
-    ]
-
-
-def get_repeater_count_for_domains(domains):
-    from .models import Repeater
-    view_kwargs = {
-        'reduce': True,
-        'include_docs': False,
-    }
-    count = 0
-    for domain in domains:
-        view_kwargs.update({
-            'startkey': [domain],
-            'endkey': [domain, {}]
-        })
-        result = Repeater.get_db().view('repeaters/repeaters', **view_kwargs).first()
-        if result:
-            count += result['value']
-    return count
